+++ CHANGELOG.md
@@ -7,6 +7,7 @@ AppEngine version, listed here to ease deployment and troubleshooting.
  * Upgraded stable Flutter analysis SDK to `3.3.5`.
  * Upgraded preview Dart analysis SDK to `2.19.0-255.2.beta`.
  * Upgraded preview Flutter analysis SDK to `3.4.0-34.1.pre`.
+ * NOTE: Removed dartdoc documentation text from search index.
 
 ## `20221004t093700-all`